Removing the MainContent component and adding everything to the MovieList
